+++ src/main/java/com/bitbank/config/SecurityConfiguration.java
@@ -3,20 +3,35 @@ package com.bitbank.config;
 import java.util.Arrays;
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
 
+import com.bitbank.services.v1.UserDetailsServiceImpl;
+
 @Configuration
 @EnableWebSecurity
+@EnableGlobalMethodSecurity(prePostEnabled = true)
 public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-	
+
+	@Autowired
+	UserDetailsServiceImpl userDetailsServiceImpl;
+
+	@Autowired
+	private AuthEntryPointJwt authEntryPointJwt;
+
 	@Bean
 	public AuthTokenFilter authenticationJwtTokenFilter() {
 		return new AuthTokenFilter();
@@ -27,14 +42,27 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 				"https://online-banking-frontend.vercel.app");
 	}
 
+	@Bean
+	public PasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+
+	@Override
+	public void configure(AuthenticationManagerBuilder authenticationManagerBuilder) throws Exception {
+		authenticationManagerBuilder.userDetailsService(userDetailsServiceImpl).passwordEncoder(passwordEncoder());
+	}
+
 	@Override
 	protected void configure(HttpSecurity http) throws Exception {
-
 		// Enable CORS and disable CSRF
 		http = http.cors().and().csrf().disable();
-		http.authorizeRequests().mvcMatchers("/api/v1/ping").permitAll()
-				.mvcMatchers("/api/v1/bank-account-transactions").permitAll().anyRequest().authenticated();
+
+		http.cors().and().csrf().disable().exceptionHandling().authenticationEntryPoint(authEntryPointJwt).and()
+				.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and().authorizeRequests()
+				.mvcMatchers("/api/v1/bank-account-transactions").permitAll().mvcMatchers("/api/v1/ping").permitAll()
+				.anyRequest().authenticated();
 		http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter.class);
+
 	}
 
 	@Bean